/branches/qt: added set of functions for size_t, added size_t flag for data generator, specialized cmnSerializeRaw for dynamic vectors to work with data generator.
git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/2012-05-14-mtsQt@3787 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisstRobot/code/robCubicSO3.cpp
@@ -79,7 +79,7 @@ robCubicSO3::SLERP( const vctQuaternionRotation3<double> &q1,
 		    double t ){
 
   double ctheta = q1.X()*q2.X() + q1.Y()*q2.Y() + q1.Z()*q2.Z() + q1.R()*q2.R();
-  double theta = acosf(ctheta);
+  double theta = acos(ctheta);
 
   // if theta = k*pi then result is not fully defined
   // we could rotate around any axis normal to qinitial or qfinal
